+++ spectest/address/address.3/address.3.go
@@ -31,22 +31,22 @@ func (m *Module) X32_good2(v0 int32) float32 {
 }
 func (m *Module) X32_good3(v0 int32) float32 {
 	t0 := v0
-	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[uint32(t0)+1:]))
+	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[int64(uint32(t0))+1:]))
 	return t1
 }
 func (m *Module) X32_good4(v0 int32) float32 {
 	t0 := v0
-	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[uint32(t0)+2:]))
+	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[int64(uint32(t0))+2:]))
 	return t1
 }
 func (m *Module) X32_good5(v0 int32) float32 {
 	t0 := v0
-	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[uint32(t0)+8:]))
+	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[int64(uint32(t0))+8:]))
 	return t1
 }
 func (m *Module) X32_bad(v0 int32) {
 	t0 := v0
-	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[uint32(t0)+4294967295:]))
+	t1 := math.Float32frombits(binary.LittleEndian.Uint32(m.memory[int64(uint32(t0))+4294967295:]))
 	_ = t1
 }
